Adding automatic tests and trying to fix finnish language
