+++ src/lang/ro/date.php
@@ -16,13 +16,13 @@ return array(
     'from_now'  => ':time de acum',
     'after'     => 'peste :time',
     'before'    => 'acum :time',
-    'year'      => 'un an|:count ani',
-    'month'     => 'o luna|:count luni',
-    'week'      => 'o saptamana|:count saptamani',
-    'day'       => 'o zi|:count zile',
-    'hour'      => 'o ora|:count ore',
+    'year'      => 'un an|:count ani|:count ani',
+    'month'     => 'o luna|:count luni|:count luni',
+    'week'      => 'o saptamana|:count saptamani|:count saptamani',
+    'day'       => 'o zi|:count zile|:count zile',
+    'hour'      => 'o ora|:count ore|:count ore',
     'minute'    => 'un minut|:count minute|:count minute',
-    'second'    => 'o secunda|:count secunde',
+    'second'    => 'o secunda|:count secunde|:count secunde',
 
     'january'   => 'ianuarie',
     'february'  => 'februarie',
