+++ tests/TranslationTest.php
@@ -72,107 +72,4 @@ class TranslationTest extends PHPUnit_Framework_TestCase {
         $this->assertSame("vor 5 Sekunden", $date->ago());
     }
 
-    public function testTranslatesMonths()
-    {
-        $months = array(
-            'january',
-            'february',
-            'march',
-            'april',
-            'may',
-            'june',
-            'july',
-            'august',
-            'september',
-            'october',
-            'november',
-            'december'
-        );
-
-        foreach ($this->languages as $language)
-        {
-            $translations = include "src/lang/$language/date.php";
-
-            foreach ($months as $month)
-            {
-                $date = new Date("1 $month");
-                $date->setLocale($language);
-
-                $this->assertTrue(isset($translations[$month]));
-                $this->assertEquals($translations[$month], $date->format('F'), "Language: $language"); // Full
-                $this->assertEquals(substr($translations[$month], 0 , 3), $date->format('M'), "Language: $language"); // Short
-            }
-        }
-    }
-
-    public function testTranslatesDays()
-    {
-        $days = array(
-            'monday',
-            'tuesday',
-            'wednesday',
-            'thursday',
-            'friday',
-            'saturday',
-            'sunday'
-        );
-
-        foreach ($this->languages as $language)
-        {
-            $translations = include "src/lang/$language/date.php";
-
-            foreach ($days as $day)
-            {
-                $date = new Date($day);
-                $date->setLocale($language);
-
-                $this->assertTrue(isset($translations[$day]));
-                $this->assertEquals($translations[$day], $date->format('l'), "Language: $language"); // Full
-                $this->assertEquals(substr($translations[$day], 0 , 3), $date->format('D'), "Language: $language"); // Short
-            }
-        }
-    }
-
-    public function testTranslatesDiffForHumans()
-    {
-        $items = array(
-            'ago',
-            'from_now',
-            'after',
-            'before',
-            'year',
-            'month',
-            'week',
-            'day',
-            'hour',
-            'minute',
-            'second'
-        );
-
-        foreach ($this->languages as $language)
-        {
-            $translations = include "src/lang/$language/date.php";
-
-            foreach ($items as $item)
-            {
-                $this->assertTrue(isset($translations[$item]), "Language: $language >> $item");
-
-                if ( ! $translations[$item])
-                {
-                    echo "\nWARNING! '$item' not set for language $language";
-                    continue;
-                }
-
-                if (in_array($item, array('ago', 'from_now', 'after', 'before')))
-                {
-                    $this->assertContains(':time', $translations[$item], "Language: $language");
-                }
-                else
-                {
-                    $this->assertContains(':count', $translations[$item], "Language: $language");
-                }
-            }
-        }
-    }
-
 }
